+++ src/Controller/ContactFormAdminController.php
@@ -11,7 +11,7 @@ use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
+use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\Routing\RouterInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
@@ -24,52 +24,20 @@ use Twig\Environment;
 
 class ContactFormAdminController
 {
-    /** @var Environment */
-    private $templatingEngine;
-    /** @var TranslatorInterface */
-    private $translator;
-    /** @var EntityManagerInterface */
-    private $entityManager;
-    /** @var SenderInterface */
-    private $mailer;
-    /** @var RouterInterface */
-    private $router;
-    /** @var FlashBagInterface */
-    private $flashBag;
-    /** @var FormFactoryInterface */
-    private $builder;
-    /** @var ContactFormMessageRepository */
-    private $contactFormMessageRepository;
-    /** @var ContactFormMessageAnswerRepository */
-    private $contactFormMessageAnswerRepository;
-    /** @var TokenStorageInterface */
-    private $token;
-
     public function __construct(
-        Environment $templatingEngine,
-        TranslatorInterface $translator,
-        EntityManagerInterface $entityManager,
-        SenderInterface $mailer,
-        RouterInterface $router,
-        FlashBagInterface $flashBag,
-        FormFactoryInterface $builder,
-        ContactFormMessageRepository $contactFormMessageRepository,
-        ContactFormMessageAnswerRepository $contactFormMessageAnswerRepository,
-        TokenStorageInterface $tokenStorage
+        private Environment $templatingEngine,
+        private TranslatorInterface $translator,
+        private EntityManagerInterface $entityManager,
+        private SenderInterface $mailer,
+        private RouterInterface $router,
+        private FormFactoryInterface $builder,
+        private ContactFormMessageRepository $contactFormMessageRepository,
+        private ContactFormMessageAnswerRepository $contactFormMessageAnswerRepository,
+        private TokenStorageInterface $tokenStorage,
     ) {
-        $this->templatingEngine = $templatingEngine;
-        $this->translator = $translator;
-        $this->entityManager = $entityManager;
-        $this->mailer = $mailer;
-        $this->router = $router;
-        $this->flashBag = $flashBag;
-        $this->builder = $builder;
-        $this->contactFormMessageRepository = $contactFormMessageRepository;
-        $this->contactFormMessageAnswerRepository = $contactFormMessageAnswerRepository;
-        $this->token = $tokenStorage;
     }
 
-    public function showMessageAction(Request $request, int $id): Response
+    public function showMessageAction(Request $request, Session $session, int $id): Response
     {
         $contactFormMessage = $this->contactFormMessageRepository->find($id);
         $contactFormMessageAnswers = $this->contactFormMessageAnswerRepository->findBy(['contactFormMessage' => $id]);
@@ -84,7 +52,7 @@ class ContactFormAdminController
             $contactFormMessage = $this->contactFormMessageRepository->find($id);
             assert($contactFormMessage instanceof ContactFormMessage);
 
-            $token = $this->token->getToken();
+            $token = $this->tokenStorage->getToken();
             assert($token && $token->getUser() instanceof AdminUser);
 
             $adminUser = $token->getUser();
@@ -97,8 +65,15 @@ class ContactFormAdminController
             $this->entityManager->persist($contactFormMessageAnswer);
             $this->entityManager->flush();
 
-            $this->mailer->send('threebrs_sylius_contact_form_answer_email', [$contactFormMessage->getEmail()], ['contact' => $contactFormMessageAnswer]);
-            $this->flashBag->add('success', $this->translator->trans('threebrs_sylius_contact_form_plugin.success'));
+            $this->mailer->send(
+                'threebrs_sylius_contact_form_answer_email',
+                [$contactFormMessage->getEmail()],
+                ['contact' => $contactFormMessageAnswer],
+            );
+            $session->getFlashBag()->add(
+                'success',
+                $this->translator->trans('threebrs_sylius_contact_form_plugin.success'),
+            );
 
             return new RedirectResponse($this->router->generate('threebrs_sylius_admin_contact_show', ['id' => $id]));
         }