+++ src/DependencyInjection/ThreeBRSSyliusContactFormExtension.php
@@ -4,17 +4,17 @@ declare(strict_types=1);
 
 namespace ThreeBRS\SyliusContactFormPlugin\DependencyInjection;
 
+use Sylius\Bundle\ResourceBundle\DependencyInjection\Extension\AbstractResourceExtension;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Loader;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 use ThreeBRS\SyliusContactFormPlugin\Model\ContactFormSettingsProvider;
 
-class ThreeBRSSyliusContactFormExtension extends Extension
+class ThreeBRSSyliusContactFormExtension extends AbstractResourceExtension
 {
     public function load(array $configs, ContainerBuilder $container): void
     {
-        $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
+        $loader = new YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.yml');
 
         $configuration = new Configuration();