+++ src/Model/ContactFormSettingsProvider.php
@@ -8,15 +8,19 @@ class ContactFormSettingsProvider implements ContactFormSettingsProviderInterfac
 {
     /** @var bool */
     private $nameRequired;
+
     /** @var bool */
     private $phoneRequired;
+
     /** @var bool */
     private $sendManager;
+
     /** @var bool */
     private $sendCustomer;
 
-    public function __construct(array $config)
-    {
+    public function __construct(
+        array $config,
+    ) {
         $this->nameRequired = $config['name_required'];
         $this->phoneRequired = $config['phone_required'];
         $this->sendManager = $config['send_manager_mail'];