+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v1.1.0 (2023-01-05)
+
+#### Details
+
+- Add support for Sylius 1.11|1.12, Symfony ^5.4|^6.0, PHP ^8.0
+- Drop support for Sylius <= 1.10 and consequentially for Symfony <= 5.3, <= PHP 7.4
+
 ## v1.0.0 (2021-11-29)
 
 ### Details
@@ -9,4 +16,4 @@
 - Change table name from `mangoweb_contact_form` to `threebrs_sylius_contact_form`
 - Change table name from `mangoweb_contact_form_message_answer` to `threebrs_sylius_contact_form_message_answer`
 - Change config name from `mango_sylius_contact_form` to `three_brs_sylius_contact_form`
-- *BC break: Version 1.0.0 is NOT compatible with previous versions due to namespace change*
+- ⚠️ *BC break: Version 1.0.0 is NOT compatible with previous versions due to **namespace change***